patch -Z not working if timestamp contains seconds > 59.5
Hi,

this is from a recent discussion in the cygwin mailing list:
(http://cygwin.com/ml/cygwin/2006-04/msg00699.html)

The patch options -Z and -T do not set filetime if the seconds part
of the timestamp is greater 59.5. This is because maketime() does
not handle denormalized times and returns an error on tm_sec = 60.

The bug could also be reproduced with patch 2.5.9 on Linux.

It might be better to ignore the fractional part of the second at all.
This would make timestamps set by patch identical to those set by tar.

[agruen: Christian's second version of the patch.]