Remember which reject files have been created
When a patch references a file more than once, patch may override
reject files it had just created.

Avoid this by first checking whether we have created reject files
ourselves, and append to such files instead of overwriting them in
that case.  (This is similar to the backup file case except that
in the backup case, we must skip creating another backup instead
of appending.)