Disable trailing CR stripping with --binary
Patch recognizes patches that have had their line endings converted
to CRLF and automatically strips CRs in that case.  This generally is
the desired behavior, but if both the patch and the file(s) being patched have
CRLF line endings, patches that would apply without this heuristic will fail.

Modify the --binary option to disable the trailing CR stripping heuristic.