Be more careful about (time_t) -1
* src/patch.c (main):
* src/pch.c (intuit_diff_type):
* src/util.c (fetchname):
Mark an invalid timespec with both .tv_sec = (time_t) -1
and with .tv_nsec = -1.  This is more reliable in case
time_t is unsigned and narrower than int, in which case
(time_t) -1 != -1.  It’s also more reliable in the unusual case
where (time_t) -1 is a valid timestamp.  All uses changed.